backend/requirements.txt
fastapi==0.115.0
uvicorn[standard]==0.32.0
python-dotenv==1.0.1
pydantic==2.9.2
pydantic-settings==2.6.0
google-generativeai==0.8.3
google-cloud-firestore==2.19.0
python-multipart==0.0.12
websockets==13.1
aiofiles==24.1.0
httpx==0.27.2
python-jose[cryptography]==3.3.0
passlib[bcrypt]==1.7.4
pypdf2==3.0.1
beautifulsoup4==4.14.2
youtube-transcript-api==1.2.4
ffmpeg-python==0.2.0
python-magic==0.4.27
requests==2.32.5
aiohttp==3.13.0
